C(TestPagination), U(TestService): them phan trang cho bai kiem tra va them so lieu thong ke

+++ src/main/java/com/englishweb/h2t_backside/dto/test/SubmitTestDTO.java
@@ -16,7 +16,7 @@ public class SubmitTestDTO extends AbstractBaseDTO {
 
     @NotNull(message = "Score cannot be null")
     @Min(value = 0, message = "Score must be at least 0")
-    private Integer score;
+    private Double score;
 
     @NotNull(message = "User id cannot be null")
     private Long userId;
@@ -25,4 +25,5 @@ public class SubmitTestDTO extends AbstractBaseDTO {
     private Long testId;
 
     private String comment;
+
 }
